requirements-lock.txt
alabaster==1.0.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==4.0.2
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
Authlib==1.6.5
babel==2.17.0
bandit==1.9.1
beautifulsoup4==4.14.2
black==25.11.0
bleach==6.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
coverage==7.11.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distlib==0.4.0
docutils==0.21.2
dparse==0.6.4
execnet==2.1.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flake8==7.3.0
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
GPUtil==1.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
identify==2.6.15
idna==3.11
ImageHash==4.3.2
imagesize==1.4.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
line_profiler==5.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==4.1.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.4
ml_dtypes==0.5.4
mpmath==1.3.0
mypy==1.18.2
mypy_extensions==1.1.0
myst-parser==4.0.1
narwhals==2.11.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.2.6
nvidia-ml-py==13.580.82
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.1.3
pure_eval==0.2.3
py-spy==0.4.1
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.3
pyparsing==3.2.5
pyreadline3==3.5.4
pytest==9.0.1
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytokens==0.3.0
pytz==2025.2
PyWavelets==1.9.0
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
roman-numerals-py==3.1.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safety==3.7.0
safety-schemas==0.0.16
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.2.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
stevedore==5.5.0
streamlit==1.51.0
stringzilla==4.2.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
types-setuptools==80.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
virtualenv==20.35.4
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
